arg_sizes need to be identical types, so use uint64_t everywhere

+++ src/HexagonOffload.cpp
@@ -179,8 +179,7 @@ class InjectHexagonRpc : public IRMutator {
             std::vector<Expr> arg_flags;
 
             for (const auto& i : c.buffers) {
-                // sizeof(pointer) will always fit into int32
-                arg_sizes.push_back(Expr((int32_t)sizeof(buffer_t*)));
+                arg_sizes.push_back(Expr((uint64_t) sizeof(buffer_t*)));
                 arg_ptrs.push_back(Variable::make(type_of<buffer_t *>(), i.first + ".buffer"));
                 int flags = 0;
                 if (i.second.read) flags |= 0x1;
@@ -192,13 +191,13 @@ class InjectHexagonRpc : public IRMutator {
                 Expr arg_ptr = Call::make(type_of<void *>(), Call::make_struct, {arg}, Call::Intrinsic);
 
                 // sizeof(scalar-type) will always fit into int32
-                arg_sizes.push_back(Expr((int32_t)i.second.bytes()));
+                arg_sizes.push_back(Expr((uint64_t) i.second.bytes()));
                 arg_ptrs.push_back(arg_ptr);
                 arg_flags.push_back(0x0);
             }
 
             // The argument list is terminated with an argument of size 0.
-            arg_sizes.push_back(Expr((int32_t)0));
+            arg_sizes.push_back(Expr((uint64_t) 0));
 
             std::string pipeline_name = hex_name + "_argv";
             std::vector<Expr> params;
